Add files via upload
The code is not finalised but you can see the structure.  The two lines at the bottom in main() are what you will need to include in your program to create an instance of my Helmet class and begin receiving data.  I will need two argument which will need to be input by the user:
1. The com port number in a string e.g. 'COM4'
2. The name of the person in a string e.g. 'Jesus'

The functions without the double underscore at the start of the name are available for you to use, the others are private to the function in my Helmet class.
I had made two simple functions for you:
1. display_name() which returns the name of the player using the Helmet.  This can be called when the user clicks on a button to open the players profile (which should display their name and call the graph for their helmet readings which you can speak to Marta about).
2. check_flag() This is something you should be calling constantly in your loop and if the return type is true than the threshold values have been exceeded and a warning should be displayed that this player is likely to have a concussion (again you can call the display_name() to show which player it is.